disable blst patch
