+++ Cargo.toml
@@ -25,7 +25,7 @@ aws-sdk-kms = "1.58.0"
 bincode = "1.3.3"
 bytemuck = "1.12"
 clap = { version = "4.5.21", features = ["derive", "env"] }
-c-kzg = "2.1.0"
+c-kzg = "=2.1.0"
 foundry-compilers = "0.11.0"
 gcloud-sdk = { version = "0.27.0", features = ["google-cloud-kms-v1"] }
 hashbrown = { version = "0.15.0", features = ["rayon"] }
