+++ bin/client/src/proving.rs
@@ -100,7 +100,7 @@ where
     P: PreimageOracleClient + Send + Sync + Debug + Clone + 'static,
     H: HintWriterClient + Send + Sync + Debug + Clone + 'static,
 {
-    // preload all data natively into a hashmap
+    // preload all data into the vec oracle
     let (_, execution_cache) = split_executions(stitched_executions.clone());
     info!(
         "Running vec witgen client with {} cached executions ({} traces).",
@@ -131,6 +131,9 @@ where
 
     let execution_trace =
         core::mem::replace(&mut witness_vec.stitched_executions, stitched_executions);
+
+    // sanity check kzg proofs
+    let _ = kailua_common::blobs::PreloadedBlobProvider::from(witness_vec.blobs_witness.clone());
 
     // check if we can prove this workload
     let (main_witness_size, witness_size) = sum_witness_size(&witness_vec);
