+++ justfile
@@ -17,6 +17,12 @@ clippy:
 
   cargo clippy --manifest-path build/risczero/fpvm/Cargo.toml --locked --workspace --all --all-targets -- -D warnings
 
+coverage:
+  cargo +nightly llvm-cov -p kailua-common --branch
+
+coverage-open:
+  cargo +nightly llvm-cov -p kailua-common --branch --open
+
 devnet-fetch:
   git clone --depth 1 --branch v1.9.1 --recursive https://github.com/ethereum-optimism/optimism.git
